:truck: Move files according to the project layout defined in #38.
+ References to the old directory layout are renamed or deleted
