+++ song2txt/__main__.py
@@ -1,0 +1,3 @@
+from .pitch_cli import main
+
+main()
