--- song2txt/uspitch/pyproject.toml
@@ -1,19 +1,0 @@
-[tool.poetry]
-name = "uspitch"
-version = "0.1.0"
-description = "Command-line utility to estimate pitches for a UltraStar txt file from the audio of the song"
-authors = ["dipzza <dipzza@disroot.org>"]
-license = "AGPLv3"
-
-[tool.poetry.dependencies]
-python = ">=3.7, <3.11"
-charset-normalizer = "^2.1.0"
-crepe = "^0.0.12"
-pydub = "^0.25.1"
-tensorflow = "^2.9.1"
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
