--- song2txt/uspitch/uspitch/__main__.py
@@ -1,3 +1,0 @@
-from .cli import main
-
-main()
