+++ tests/test_pitch_cli.py
@@ -1,7 +1,7 @@
 from unittest import mock
 import sys
 
-from song2txt.uspitch.uspitch.cli import main
+from song2txt.pitch_cli import main
 
 import pytest
 import numpy as np
@@ -26,25 +26,25 @@ def crepe_value():
     return [0, frequency, confidence, 0]
 
 
-@mock.patch('song2txt.uspitch.uspitch.cli.crepe')
-@mock.patch('song2txt.uspitch.uspitch.cli.txt')
+@mock.patch('song2txt.pitch_cli.crepe')
+@mock.patch('song2txt.pitch_cli.txt')
 def test_no_required_arguments(*mocks):
     with pytest.raises(SystemExit, match='2'):
         main()
 
 
-@mock.patch('song2txt.uspitch.uspitch.cli.crepe')
-@mock.patch('song2txt.uspitch.uspitch.cli.txt')
+@mock.patch('song2txt.pitch_cli.crepe')
+@mock.patch('song2txt.pitch_cli.txt')
 def test_incorrect_type(*mocks):
     sys.argv = ['cli.py', 'path', '--confidence', 'str']
     with pytest.raises(SystemExit, match='2'):
         main()
 
 
-@mock.patch('song2txt.uspitch.uspitch.cli.open_audio', return_value=[0, 0])
-@mock.patch('song2txt.uspitch.uspitch.cli.crepe')
-@mock.patch('song2txt.uspitch.uspitch.cli.txt.write_file')
-@mock.patch('song2txt.uspitch.uspitch.cli.txt.read_file')
+@mock.patch('song2txt.pitch_cli.open_audio', return_value=[0, 0])
+@mock.patch('song2txt.pitch_cli.crepe')
+@mock.patch('song2txt.pitch_cli.txt.write_file')
+@mock.patch('song2txt.pitch_cli.txt.read_file')
 def test_default_args(m_read, m_write, m_crepe, _, read_value, crepe_value):
     sys.argv = ['cli.py', '../filepath.txt']
 
@@ -59,10 +59,10 @@ def test_default_args(m_read, m_write, m_crepe, _, read_value, crepe_value):
     assert np.array_equal(m_write.call_args[0][2], np.array([60, -1, 60]))
 
 
-@mock.patch('song2txt.uspitch.uspitch.cli.open_audio', return_value=[0, 0])
-@mock.patch('song2txt.uspitch.uspitch.cli.crepe')
-@mock.patch('song2txt.uspitch.uspitch.cli.txt.write_file')
-@mock.patch('song2txt.uspitch.uspitch.cli.txt.read_file')
+@mock.patch('song2txt.pitch_cli.open_audio', return_value=[0, 0])
+@mock.patch('song2txt.pitch_cli.crepe')
+@mock.patch('song2txt.pitch_cli.txt.write_file')
+@mock.patch('song2txt.pitch_cli.txt.read_file')
 def test_custom_output(m_read, m_write, m_crepe, _, read_value, crepe_value):
     sys.argv = ['cli.py', 'path', '-o', 'custom_output.txt']
